fix: split analysis helpers for client
