+++ src/lib/analysisHelpers.ts
@@ -1,0 +1,17 @@
+import type { ViolationReport } from "./openai";
+
+export function mergeAnalysis(
+  base: ViolationReport | null | undefined,
+  overrides: Partial<ViolationReport> | null | undefined,
+): ViolationReport | null {
+  if (!base) return null;
+  if (!overrides) return base;
+  return {
+    ...base,
+    ...overrides,
+    vehicle: {
+      ...base.vehicle,
+      ...(overrides.vehicle ?? {}),
+    },
+  };
+}
